extract components from FieldLabel
