+++ src/FieldLabel/FieldLabel.tsx
@@ -1,29 +1,10 @@
 import React from "react";
-import styled from "styled-components";
-import { space, color } from "styled-system";
-import PropTypes from "prop-types";
 import { Box } from "../Box";
-import RequirementText from "./RequirementText";
 import HelpText from "./HelpText";
+import Label from "./Label";
+import LabelText from "./LabelText";
+import RequirementText from "./RequirementText";
 import { FieldLabelProps, FieldLabelDefaultProps } from "./FieldLabel.type";
-
-const Label = styled.label(space, color, () => ({
-  display: "inline-block",
-}));
-
-Label.propTypes = {
-  color: PropTypes.string,
-};
-
-Label.defaultProps = {
-  color: "black",
-};
-
-const LabelText = styled.span(({ theme }) => ({
-  fontSize: theme.fontSizes.small,
-  fontWeight: theme.fontWeights.bold,
-  lineHeight: theme.lineHeights.smallTextCompressed,
-}));
 
 const FieldLabel = ({
   labelText,
