+++ src/FieldLabel/LabelText.tsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+const LabelText = styled.span(({ theme }) => ({
+  fontSize: theme.fontSizes.small,
+  fontWeight: theme.fontWeights.bold,
+  lineHeight: theme.lineHeights.smallTextCompressed,
+}));
+
+export default LabelText;
+
